requirements.txt
paho-mqtt>=1.6.1
pyserial>=3.5
PyYAML>=6.0
